update core + plugins

+++ wp-content/plugins/all-in-one-wp-security-and-firewall/classes/wp-security-user-login.php
@@ -324,6 +324,7 @@ class AIOWPSecurity_User_Login {
 		$lock_minutes = $this->get_dynamic_lockout_time_length();
 		$ip = AIOWPSecurity_Utility_IP::get_user_ip_address(); //Get the IP address of user
 		if (empty($ip)) return;
+		$username = sanitize_user($username, true); // Sanitize as a username
 		$ip_range = AIOWPSecurity_Utility_IP::get_sanitized_ip_range($ip); //Get the IP range of the current user
 		$user = is_email($username) ? get_user_by('email', $username) : get_user_by('login', $username); //Returns WP_User object if exists
 		$ip_range = apply_filters('aiowps_before_lockdown', $ip_range);
@@ -399,8 +400,11 @@ class AIOWPSecurity_User_Login {
 				$email_msg = __('User login lockout events had occurred due to too many failed login attempts or invalid username:', 'all-in-one-wp-security-and-firewall')."\n\n";
 			
 				foreach ($lockout_ips_list as $lockout_ip) {
+					// Prevent email clients from auto-linking URLs in the username.
+					$lockout_ip_username = str_replace(array('.', '@'), array('[.]', '[at]'), $lockout_ip['username']);
+					
 					/* translators: %s: User name. */
-					$email_msg .= sprintf(__('Username: %s', 'all-in-one-wp-security-and-firewall'), $lockout_ip['username']) . "\n";
+					$email_msg .= sprintf(__('Username: %s', 'all-in-one-wp-security-and-firewall'), $lockout_ip_username) . "\n";
 
 					/* translators: %s: IP Address. */
 					$email_msg .= sprintf(__('IP address: %s', 'all-in-one-wp-security-and-firewall'), $lockout_ip['ip']) . "\n";
@@ -427,8 +431,17 @@ class AIOWPSecurity_User_Login {
 				$email_msg .= __("Log into your site WordPress administration panel to see the duration of the lockout or to unlock the user.", 'all-in-one-wp-security-and-firewall') . "\n";
 
 				$email_header = '';
-				$send_mail = wp_mail($to_email_address, $subject, $email_msg, $email_header, $backtrace_filepath);
-			
+
+				$mail_data = array(
+					'to' => $to_email_address,
+					'subject' => $subject,
+					'message' => $email_msg,
+					'headers' => $email_header,
+					'attachments' => $backtrace_filepath
+				);
+
+				$send_mail = AIOWPSecurity_Reporting::notification($mail_data);
+
 				if (false === $send_mail) {
 					$ips_list = implode(', ', wp_list_pluck($lockout_ips_list, 'ip'));
 					$aio_wp_security->debug_logger->log_debug("Lockout notification email failed to send to " . implode(', ', $to_email_address) . " for IPs ".$ips_list, 4);
@@ -443,19 +456,24 @@ class AIOWPSecurity_User_Login {
 	 *
 	 * @global type $wpdb
 	 * @global AIO_WP_Security $aio_wp_security
+	 *
 	 * @param type $ip_range
+	 * @param int  $user_id
+	 *
 	 * @return string or false on failure
 	 */
-	public static function generate_unlock_request_link($ip_range) {
+	public static function generate_unlock_request_link($ip_range, $user_id) {
 		//Get the locked user row from locout table
 		global $wpdb, $aio_wp_security;
 		$unlock_link = '';
 		$lockout_table_name = AIOWPSEC_TBL_LOGIN_LOCKOUT;
 		$secret_rand_key = (md5(uniqid(wp_rand(), true)));
 		// phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared, WordPress.DB.DirectDatabaseQuery -- PCP wanting. Ignore.
-		$res = $wpdb->query($wpdb->prepare("UPDATE $lockout_table_name SET unlock_key = %s WHERE released > UNIX_TIMESTAMP() AND failed_login_ip LIKE %s", $secret_rand_key,  "%" . esc_sql($ip_range) . "%"));
-		if (null == $res) {
-			$aio_wp_security->debug_logger->log_debug("No locked user found with IP range ".$ip_range, 4);
+		$res = $wpdb->query($wpdb->prepare("UPDATE $lockout_table_name SET unlock_key = %s WHERE released > UNIX_TIMESTAMP() AND failed_login_ip LIKE %s AND user_id = %d", $secret_rand_key,  "%" . esc_sql($ip_range) . "%", $user_id));
+		if (!$res) { // 0 or false will be returned by $wpdb->query
+			$locked_user_data = get_userdata($user_id);
+			$locked_user_email = $locked_user_data ? $locked_user_data->user_email : '';
+			$aio_wp_security->debug_logger->log_debug("No locked entry was found in the database matching your IP address range " . $ip_range . " and user account " . $locked_user_email, 4);
 			return false;
 		} else {
 			// Check if unlock request or submitted from a WooCommerce account login page
@@ -544,9 +562,16 @@ class AIOWPSecurity_User_Login {
 		$subject = '['.network_site_url().'] '. __('Unlock request notification', 'all-in-one-wp-security-and-firewall');
 		/* translators: 1: Email 2: Link */
 		$email_msg = sprintf(__('You have requested for the account with email address %s to be unlocked.', 'all-in-one-wp-security-and-firewall') . ' ' . __('Please press the link below to unlock your account:', 'all-in-one-wp-security-and-firewall'), $email) . "\n" . sprintf(__('Unlock link: %s', 'all-in-one-wp-security-and-firewall'), $unlock_link) . "\n\n" . __('After pressing the above link you will be able to login to the WordPress administration panel.', 'all-in-one-wp-security-and-firewall') . "\n";
-		
-		$sendMail = wp_mail($email, $subject, $email_msg);
-		if (false === $sendMail) {
+
+		$mail_data = array(
+			'to' => $email,
+			'subject' => $subject,
+			'message' => $email_msg,
+		);
+
+		$send_mail = AIOWPSecurity_Reporting::notification($mail_data);
+
+		if (false === $send_mail) {
 			$aio_wp_security->debug_logger->log_debug("Unlock Request Notification email failed to send to " . $email, 4);
 		}
 	}
@@ -577,7 +602,7 @@ class AIOWPSecurity_User_Login {
 				$logout_time_interval_val_seconds = $logout_time_interval_value * 60;
 				if ($diff > $logout_time_interval_val_seconds) {
 					$aio_wp_security->debug_logger->log_debug("Force Logout - This user logged in more than (".$logout_time_interval_value.") minutes ago. Doing a force log out for the user with username: ".$current_user->user_login);
-					$this->wp_logout_action_handler($user_id); //this will register the logout time/date in the logout_date column
+					$this->wp_logout_action_handler($user_id, true); //this will register the logout time/date in the logout_date column
 
 
 					$curr_page_url = AIOWPSecurity_Utility::get_current_page_url();
@@ -691,7 +716,9 @@ class AIOWPSecurity_User_Login {
 			return;
 		}
 
-		$this->delete_logged_in_user($user->ID);
+		if (empty(get_user_meta($user_id, 'session_tokens', true))) {
+			$this->delete_logged_in_user($user->ID);
+		}
 
 		if (is_super_admin($user->ID)) {
 			$logging_out_of_correct_site = true;
@@ -772,7 +799,7 @@ class AIOWPSecurity_User_Login {
 		$unlock_secret_string = $aio_wp_security->configs->get_value('aiowps_unlock_request_secret_key');
 		$current_time = time();
 		$enc_result = base64_encode($current_time.$unlock_secret_string);
-		$unlock_request_form .= '<form method="post" action=""><div style="padding-bottom:10px;"><input type="hidden" name="aiowps-unlock-string-info" id="aiowps-unlock-string-info" value="'.$enc_result.'" />';
+		$unlock_request_form .= '<form method="post" action="">'.wp_nonce_field('aios-unlock-nonce', '_wpnonce', true, false).'<div style="padding-bottom:10px;"><input type="hidden" name="aiowps-unlock-string-info" id="aiowps-unlock-string-info" value="'.$enc_result.'" />';
 		$unlock_request_form .= '<input type="hidden" name="aiowps-unlock-temp-string" id="aiowps-unlock-temp-string" value="'.$current_time.'" />';
 		// phpcs:ignore WordPress.Security.NonceVerification.Missing -- No nonce.
 		if (isset($_POST['woocommerce-login-nonce'])) {
@@ -936,7 +963,7 @@ class AIOWPSecurity_User_Login {
 		if (empty($auth_cookie)) return; //check if auth cookie is empty, meaning login was not successful
 		$expiration = $expire > 0 ? $expire : $expiration;
 
-		if (is_multisite() && !is_super_admin()) {
+		if (is_multisite() && !is_super_admin($user_id)) {
 			$user_blog = get_active_blog_for_user($user_id);
 			switch_to_blog($user_blog->blog_id); // switch to user blog incase they try to log in from wrong subsite
